Rename modules using PEP8 convention
