+++ masc/constants.py
