+++ masc/custom.py
@@ -1,4 +1,4 @@
-from masc.CMS import CMS
+from masc.cms import CMS
 
 
 class Custom(CMS):
