+++ masc/masc_entry.py
